Fix realtimestats fee and margin (#2622)
* Handle margin and fee in both realtimestats functions

* Add tests

+++ server/dash/features/realtimestats/serializers.py
@@ -10,4 +10,4 @@ class AdGroupRealtimeStatsSerializer(serializers.Serializer):
 
 class AdGroupSourcesRealtimeStatsSerializer(serializers.Serializer):
     source = restapi.fields.PlainCharField()
-    spend = serializers.FloatField()
+    spend = serializers.DecimalField(max_digits=20, decimal_places=2)
